updating readme, moving signals to seperate file, removing redundant serializers
