+++ .idea/dictionaries/ahmad.xml
@@ -1,0 +1,8 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="ahmad">
+    <words>
+      <w>genertaed</w>
+      <w>updtae</w>
+    </words>
+  </dictionary>
+</component>
