+++ delivery_hero/swagger_schema.py
@@ -7,6 +7,8 @@ from rest_framework.views import APIView
 from rest_framework_swagger import renderers
 
 
+# overriding swagger to ensure that swagger also shows protected endpoints.
+# These endpoints will still need authentication credentials to work
 def get_swagger_view(title=None, url=None, patterns=None, urlconf=None):
     """
     Returns schema view which renders Swagger/OpenAPI.
