+++ requirements.txt
@@ -2,7 +2,7 @@ certifi==2018.1.18
 chardet==3.0.4
 coreapi==2.3.3
 coreschema==0.0.4
-Django==2.0.4
+Django==1.11.7
 django-rest-swagger>=2.1.2
 djangorestframework==3.7.7
 idna==2.6
