+++ restaurant/apps.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig
+
+
+class RestaurantConfig(AppConfig):
+    name = 'restaurant'
+
+    def ready(self):
+        import restaurant.signals
