+++ restaurant/signals.py
@@ -1,18 +1,11 @@
-from django.db import models
-from django.contrib.auth.models import User
 from django.conf import settings
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 from rest_framework.authtoken.models import Token
 
 
+# create an auth token for every new user
 @receiver(post_save, sender=settings.AUTH_USER_MODEL)
 def create_auth_token(sender, instance=None, created=False, **kwargs):
     if created:
         Token.objects.create(user=instance)
-
-
-class Restaurant(models.Model):
-    name = models.TextField()
-    opens_at = models.TimeField()
-    closes_at = models.TimeField()
